+++ modules/skparagraph/include/DartTypes.h
@@ -62,6 +62,7 @@ struct PositionWithAffinity {
     int32_t position;
     Affinity affinity;
 
+    PositionWithAffinity() : position(0), affinity(kDownstream) {}
     PositionWithAffinity(int32_t p, Affinity a) : position(p), affinity(a) {}
 };